refactor: Partially isolate map and table logic
