+++ scripts/modules/map.mjs
@@ -34,10 +34,8 @@ function pointToEllipse(row) {
   return ellipse;
 }
 
-function popup(row) {
-  const columns = myFunctionHolder.getCraterTableColumns();
-
-  const lines = columns
+function popup(row, columnMap) {
+  const lines = columnMap
     .map(({ label, format }) => {
       try {
         const value = format(row);
@@ -54,70 +52,9 @@ function popup(row) {
   return popupText;
 }
 
-let pageSize = 25;
-let currentPage = 0;
-
 myFunctionHolder.allDim = {};
 
-myFunctionHolder.displayPage = function (pageIndex) {
-  const totalPages = Math.ceil(myFunctionHolder.totalRows / pageSize);
-  if (pageIndex >= totalPages) return; // Don't go beyond last page
-  if (pageIndex < 0) return; // Don't go before first page
-  const offset = pageIndex * pageSize;
-  const pageData = myFunctionHolder.enrichedData.slice(
-    offset,
-    offset + pageSize,
-  ); // ✅
-
-  myFunctionHolder.allDim.filterFunction((d) => pageData.includes(d)); // ✅ filter manually
-
-  dc.redrawAll();
-  myFunctionHolder.updatePageInfo();
-};
-
-myFunctionHolder.updatePageInfo = function () {
-  const offset = currentPage * pageSize;
-  const pageText = `Showing rows ${offset + 1}–${offset + pageSize} out of ${myFunctionHolder.totalRows}`;
-  document.getElementById("page-info").textContent = pageText;
-
-  const totalPages = Math.ceil(myFunctionHolder.totalRows / pageSize);
-  document.getElementById("next").disabled = currentPage + 1 >= totalPages;
-  document.getElementById("prev").disabled = currentPage === 0;
-};
 myFunctionHolder.ellipseMap = {};
-
-// Table formatting
-myFunctionHolder.getCraterTableColumns = function () {
-  return [
-    { label: "Crater ID", format: (d) => d.CRATER_ID },
-    { label: "Lat", format: (d) => d.LAT_ELLI_IMG.toFixed(3) },
-    { label: "Lon", format: (d) => d.LON_ELLI_IMG.toFixed(3) },
-    { label: "Diameter (km)", format: (d) => d.DIAM_CIRC_IMG.toFixed(2) },
-    { label: "Eccentricity", format: (d) => d.DIAM_ELLI_ECCEN_IMG.toFixed(3) },
-    { label: "Ellipticity", format: (d) => d.DIAM_ELLI_ELLIP_IMG.toFixed(3) },
-    {
-      label: "Angle (from North)",
-      format: (d) => d.DIAM_ELLI_ANGLE_IMG.toFixed(1),
-    },
-
-    { label: "Ejecta Layers", format: (d) => d.LAY_NUMBER },
-    { label: "Ejecta Class", format: (d) => d.LAY_MORPH1 }, // @Aidan need a lookup table for this.
-    { label: "Ejecta Texture", format: (d) => d.LAY_MORPH2 }, // @Aidan need a lookup table for this.
-    { label: "Ejecta Shape", format: (d) => d.LAY_MORPH3 },
-    { label: "Ejecta Notes", format: (d) => d.LAY_NOTES },
-
-    { label: "Crater Class", format: (d) => d.INT_MORPH1 }, // @Aidan need a lookup table for this.
-    { label: "Wall Morph", format: (d) => d.INT_MORPH2 },
-    { label: "Floor Morph", format: (d) => d.INT_MORPH3 },
-
-    { label: "Confidence", format: (d) => d.CONF },
-    { label: "Notes", format: (d) => d.NOTES },
-
-    { label: "Rim Degradation", format: (d) => d.DEG_RIM },
-    { label: "Ejecta Degradation", format: (d) => d.DEG_EJC },
-    { label: "Floor Degradation", format: (d) => d.DEG_FLR },
-  ];
-};
 
 function configureMap(id) {
   let leafletMap = L.map(id).setView([0, 0], 2);
@@ -138,10 +75,20 @@ function configureMap(id) {
   return leafletMap;
 }
 
+function drawEllipse(layer, row, columnMap) {
+  const ellipse = pointToEllipse(row);
+  const popupText = popup(row, columnMap);
+
+  ellipse.bindPopup(popupText);
+  layer.addLayer(ellipse);
+
+  return ellipse;
+}
+
 export {
   configureMap,
-  pointToEllipse,
-  popup,
   myFunctionHolder,
+  setDefaultEllipseStyle,
   setSelectedEllipseStyle,
+  drawEllipse,
 };
